refactor: Main 화면 Home Fragment로 전환 및 main 모듈 card에 통합 (#121)
- :feature:main 모듈 제거
- 메인 화면을 ComposeView 기반 HomeFragment로 전환
- 리소스 및 유틸 클래스 이동 (:feature:main → :feature:card)
- 의존성 추가 (coil-gif, fragment-ktx)
- 사용하지 않는 코드 정리
