+++ app/build.gradle.kts
@@ -43,7 +43,6 @@ android {
 dependencies {
 
     implementation(project(":feature:card"))
-    implementation(project(":feature:main"))
     implementation(project(":core:data"))
 
     // Hilt
