+++ feature/card/src/main/java/com/zcard/home/HomeFragment.kt
@@ -1,0 +1,35 @@
+package com.zcard.home
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.compose.ui.platform.ComposeView
+import androidx.compose.ui.platform.ViewCompositionStrategy
+import androidx.fragment.app.Fragment
+import com.zcard.card.R
+import com.zcard.card.cardeditor.CardEditorFragment
+import com.zcard.designsystem.theme.ZCardTheme
+
+class HomeFragment : Fragment() {
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        return ComposeView(requireContext()).apply {
+            setViewCompositionStrategy(ViewCompositionStrategy.DisposeOnViewTreeLifecycleDestroyed)
+            setContent {
+                ZCardTheme {
+                    MainScreen {
+                        parentFragmentManager.beginTransaction()
+                            .replace(R.id.fragment_container, CardEditorFragment.newInstance(it))
+                            .addToBackStack(null)
+                            .commit()
+                    }
+                }
+            }
+        }
+    }
+}
