+++ feature/card/src/main/res/drawable/gif_happy_2026.gif
