+++ feature/card/src/main/res/drawable/ic_card_star.xml
