+++ gradle/libs.versions.toml
@@ -1,5 +1,6 @@
 [versions]
 agp = "8.13.0"
+androidxFragmentFragmentKtx = "1.8.9"
 coil = "2.7.0"
 hiltNavigationCompose = "1.3.0"
 kotlin = "2.0.21"
@@ -35,6 +36,7 @@ activity = "1.12.1"
 [libraries]
 androidx-compose-material-icons-extended = { module = "androidx.compose.material:material-icons-extended" }
 androidx-core-ktx = { group = "androidx.core", name = "core-ktx", version.ref = "coreKtx" }
+androidx-fragment-fragment-ktx = { module = "androidx.fragment:fragment-ktx", version.ref = "androidxFragmentFragmentKtx" }
 coil-compose = { module = "io.coil-kt:coil-compose", version.ref = "coil" }
 coil-gif = { module = "io.coil-kt:coil-gif", version.ref = "coil" }
 junit = { group = "junit", name = "junit", version.ref = "junit" }
